refactor(navigation): decompose giant Navigation component into smaller sub-components
